+++ src/main/java/insane96mcp/insanesurvivaloverhaul/setup/ISORegistries.java
@@ -8,6 +8,7 @@ import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.village.poi.PoiType;
+import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.neoforged.neoforge.registries.DeferredRegister;
@@ -20,6 +21,8 @@ public class ISORegistries {
     public static final List<DeferredRegister<?>> REGISTRIES = new ArrayList<>();
 
     public static final DeferredRegister<Item> ITEMS = createRegistry(BuiltInRegistries.ITEM);
+    public static final DeferredRegister<Item> MINECRAFT_ITEMS = createRegistryForNamespace(BuiltInRegistries.ITEM, "minecraft");
+    public static final DeferredRegister<ArmorMaterial> MINECRAFT_ARMOR_MATERIALS = createRegistryForNamespace(BuiltInRegistries.ARMOR_MATERIAL, "minecraft");
     public static final DeferredRegister<Block> BLOCKS = createRegistry(BuiltInRegistries.BLOCK);
     public static final DeferredRegister<Attribute> ATTRIBUTES = createRegistry(BuiltInRegistries.ATTRIBUTE);
     public static final DeferredRegister<SoundEvent> SOUND_EVENTS = createRegistry(BuiltInRegistries.SOUND_EVENT);
@@ -30,6 +33,12 @@ public class ISORegistries {
 
     private static <R> DeferredRegister<R> createRegistry(Registry<R> registry) {
         DeferredRegister<R> register = DeferredRegister.create(registry, InsaneSO.MOD_ID);
+        REGISTRIES.add(register);
+        return register;
+    }
+
+    private static <R> DeferredRegister<R> createRegistryForNamespace(Registry<R> registry, String namespace) {
+        DeferredRegister<R> register = DeferredRegister.create(registry, namespace);
         REGISTRIES.add(register);
         return register;
     }