+++ changelog.md
@@ -3,8 +3,10 @@
   * Not sure if and when Villager feature will come back.
 * Added back Misc features: Packs, Nerfs, Tweaks, Potions and Effects, Low Fish
   * Potion and Effects now enables a data pack that changes potions stack sizes
-* Added back Items features: Stack Sizes, Name Tags, Disabled Items and Ecologic Wood
+* Added back Items features: Stack Sizes, Name Tags, Disabled Items, Stone Tools, Copper equipment gone and Ecologic Wood
   * Items and blocks (not food) now stack up to 99 (max stack size is increased by ~54.7%)
+  * Stone tools gone now also replaces items in entities equipment slots
+  * Copper items are now registered in the minecraft namespace to be able to (maybe) update a modded world to newer versions
 * Added back World features: Fluids
 * Mobs buffs data pack has been merged in Enhanced AI's MPR Integration data pack